Genetic marker set tests based on the covariance test

Genetic marker set tests based on the covariance statistics for a set of genetic markers.

cvat(fit = NULL, s = NULL, g = NULL, W = NULL, sets = NULL,
  nperm = 100)

Arguments

fit
is the fit object obtained from a linear mixed model fit using the greml function

s
vector (or list) of single marker effects obtained from a linear mixed model fit (GBLUP of GFBLUP)

g
vector (or list) of genetic effects obtained from a linear mixed model fit (GBLUP of GFBLUP)

W
matrix of centered and scaled genotypes (n x m)

sets
list of marker sets corresponding to columns in W

nperm
number of permutations

Value

Returns a dataframe including

setT

covariance test statistics

nset

number of markers in the set

p

p-value

Details

The covariance test statistic is derived from a GBLUP (or GFBLUP) model fit. It is a measure of covariance between the total genomic effect for all markers and the genomic effect for the genetic markers in the genomic feature. It also relates to the explained sums of squares for the genetic markers. The distribution of this test statistic under the null hypothesis is difficult to describe in terms of exact or approximate distributions, and an empirical distribution is required.

Examples

# Simulate data
W <- matrix(rnorm(20000000), ncol = 10000)
        colnames(W) <- as.character(1:ncol(W))
        rownames(W) <- as.character(1:nrow(W))
y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))

# Create model
data <- data.frame(y = y, mu = 1)
fm <- y ~ 0 + mu
X <- model.matrix(fm, data = data)

# Create framework for lists
setsGB <- list(A = colnames(W)) # gblup model
setsGF <- list(C1 = colnames(W)[1:1000], C2 = colnames(W)[1001:2000], C3 = colnames(W)[2000:10000]) # gfblup model
setsGT <- list(C1 = colnames(W)[1:10], C2 = colnames(W)[1001:1010], C3 = colnames(W)[1:10000]) # true model

# Compute G
G <- computeG(W = W)
GB <- lapply(setsGB, function(x) {computeG(W = W[, x])})
GF <- lapply(setsGF, function(x) {computeG(W = W[, x])})
GT <- lapply(setsGT, function(x) {computeG(W = W[, x])})

# REML analyses and multi marker association (set) test
fitGB <- greml(y = y, X = X, G = GB, verbose = TRUE)

#> [1] "Iteration:" "1"          "Theta:"     "2.25"       "2.03"      
#> [1] "Iteration:" "2"          "Theta:"     "4.38"       "3.27"      
#> [1] "Iteration:" "3"          "Theta:"     "8.12"       "4.29"      
#> [1] "Iteration:" "4"          "Theta:"     "13.27"      "4.03"      
#> [1] "Iteration:" "5"          "Theta:"     "17.33"      "2.76"      
#> [1] "Iteration:" "6"          "Theta:"     "18.44"      "2.24"      
#> [1] "Iteration:" "7"          "Theta:"     "18.5"       "2.2"       
#> [1] "Iteration:" "8"          "Theta:"     "18.5"       "2.2"       
#> [1] "Iteration:" "9"          "Theta:"     "18.5"       "2.2"       
#> [1] "Iteration:" "10"         "Theta:"     "18.5"       "2.2"       
#> [1] "Converged at Iteration:" "10"                     
#> [3] "Theta:"                  "18.5"                   
#> [5] "2.2"                    

# Use fit object as input
cvat(fit = fitGB, W = W, sets = setsGF, nperm = 1000)

#>         setT nset     p
#> C1  5970.314 1000 0.073
#> C2  5311.328 1000 0.175
#> C3 21612.345 8001 0.926

cvat(fit = fitGB, W = W, sets = setsGT, nperm = 1000)

#>         setT  nset p
#> C1  3317.364    10 0
#> C2  2790.249    10 0
#> C3 32893.993 10000 1

# Use single coefficients as input 
s <- crossprod(W / ncol(W), fitGB$Py) * fitGB$theta[1]
cvat(s = s, W = W, sets = setsGF, nperm = 1000)

#>         setT nset     p
#> C1  5970.314 1000 0.087
#> C2  5311.328 1000 0.193
#> C3 21612.345 8001 0.911

cvat(s = s, W = W, sets = setsGT, nperm = 1000)

#>         setT  nset p
#> C1  3317.364    10 0
#> C2  2790.249    10 0
#> C3 32893.993 10000 1