Genetic marker set tests based on the hyperG test

Genetic marker set tests based on the hyperG test statistics for a set of genetic markers.

hgTest(p = NULL, sets = NULL, threshold = 0.05)

Arguments

p
vector of single marker p-values (e.g. based on a t-stat)

sets
list of marker sets - names corresponds to rownames in stat

threshold
single marker p-value cut-off

Value

Returns vector of p values with length equal to the number of sets

Details

The hyperG marker set test tests a predefined set of markers (i.e. those within a particular genomic feature) for an association with the trait phenotype. Under the null hypothesis (associated markers are picked at random from the total number of tested genetic markers) it is assumed that the observed count statistic is a realization from a hypergeometric distribution. This hypothesis can be formulated and tested in a number of ways. Here we consider a test statistic based on counting the number of genetic markers in the feature that are associated to trait phenotype. A test based on the count test statistic is likely to have high power to detect association if the genomic feature harbours genetic markers with large effects.

Examples

# Simulate data
W <- matrix(rnorm(2000000), ncol = 10000)
  colnames(W) <- as.character(1:ncol(W))
y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))

# REML analyses 
data <- data.frame(y = y, mu = 1)
fm <- y ~ mu
fit <- gfm(fm = fm, W = W, sets = list(colnames(W)), data = data)

#> start algorithm at 10.407 10.407 
#> 1 sigma = 10.40698 10.40698 resid llik = -401.0601 
#> 1 adjusted sigma = 10.39093 10.39093
#> 2 sigma = 14.01627 7.73041 resid llik = -400.9869 
#> 2 adjusted sigma = 13.41037 7.396234 delta.llik = 0.07322748 
#> 3 sigma = 16.54537 5.08183 resid llik = -400.9498 
#> 3 adjusted sigma = 15.94261 4.896695 delta.llik = 0.0370784 
#> 4 sigma = 17.99955 3.320971 resid llik = -400.9377 
#> 4 adjusted sigma = 17.61664 3.250321 delta.llik = 0.01209583 
#> 5 sigma = 18.43876 2.550268 resid llik = -400.9356 
#> 5 adjusted sigma = 18.34338 2.537077 delta.llik = 0.002133997 
#> 6 sigma = 18.65455 2.251895 resid llik = -400.9353 
#> 6 adjusted sigma = 18.63645 2.249711 delta.llik = 0.0003226468 

# hyperG set test 
sets <- list(A = as.character(1:100), B = as.character(101:1000), C = as.character(1001:5000), D = as.character(5001:10000))
p <- pt(fit$s / fit$vs)

#> Error in pt(fit$s/fit$vs): argument "df" is missing, with no default

res <- hgTest(p = p, sets = sets, threshold = 0.05)

#> Error in hgTest(p = p, sets = sets, threshold = 0.05): object 'p' not found