Linear mixed model marker association test

Marker test using linear mixed model (LMM) to test for an association of single markers with a phenotype.

lmma(fit = NULL, W = NULL, m = NULL, statistic = "mastor")

Arguments

fit
list of information about linear model fit (output from greml)

W
matrix of centered and scaled genotypes (n x m)

m
is the total number of markers in W

statistic
is the single marker test statistic used. Default is the "mastor", alternatives include "gblup", "bolt-lmm" and "grammar-gamma"

Value

Returns a dataframe including

coef

single marker coefficients

se

standard error of coefficients

stat

single marker test statistic

p

p-value

Details

Linear mixed model single marker association (LMMA) statistics are based on either exact or approximate methods. Exact methods estimate variance components and effects of single markers jointly. Approximate methods estimate single marker effects conditionally.

References

Chen, W. M., & Abecasis, G. R. (2007). Family-based association tests for genomewide association scans. The American Journal of Human Genetics, 81(5), 913-926.

Loh, P. R., Tucker, G., Bulik-Sullivan, B. K., Vilhjalmsson, B. J., Finucane, H. K., Salem, R. M., ... & Patterson, N. (2015). Efficient Bayesian mixed-model analysis increases association power in large cohorts. Nature genetics, 47(3), 284-290.

Kang, H. M., Sul, J. H., Zaitlen, N. A., Kong, S. Y., Freimer, N. B., Sabatti, C., & Eskin, E. (2010). Variance component model to account for sample structure in genome-wide association studies. Nature genetics, 42(4), 348-354.

Lippert, C., Listgarten, J., Liu, Y., Kadie, C. M., Davidson, R. I., & Heckerman, D. (2011). FaST linear mixed models for genome-wide association studies. Nature methods, 8(10), 833-835.

Listgarten, J., Lippert, C., Kadie, C. M., Davidson, R. I., Eskin, E., & Heckerman, D. (2012). Improved linear mixed models for genome-wide association studies. Nature methods, 9(6), 525-526.

Listgarten, J., Lippert, C., & Heckerman, D. (2013). FaST-LMM-Select for addressing confounding from spatial structure and rare variants. Nature Genetics, 45(5), 470-471.

Lippert, C., Quon, G., Kang, E. Y., Kadie, C. M., Listgarten, J., & Heckerman, D. (2013). The benefits of selecting phenotype-specific variants for applications of mixed models in genomics. Scientific reports, 3.

Zhou, X., & Stephens, M. (2012). Genome-wide efficient mixed-model analysis for association studies. Nature genetics, 44(7), 821-824.

Svishcheva, G. R., Axenovich, T. I., Belonogova, N. M., van Duijn, C. M., & Aulchenko, Y. S. (2012). Rapid variance components-based method for whole-genome association analysis. Nature genetics, 44(10), 1166-1170.

Yang, J., Zaitlen, N. A., Goddard, M. E., Visscher, P. M., & Price, A. L. (2014). Advantages and pitfalls in the application of mixed-model association methods. Nature genetics, 46(2), 100-106.

Bulik-Sullivan, B. K., Loh, P. R., Finucane, H. K., Ripke, S., Yang, J., Patterson, N., ... & Schizophrenia Working Group of the Psychiatric Genomics Consortium. (2015). LD Score regression distinguishes confounding from polygenicity in genome-wide association studies. Nature genetics, 47(3), 291-295.

Examples

# Simulate data
W <- matrix(rnorm(20000000), ncol = 10000)
        colnames(W) <- as.character(1:ncol(W))
        rownames(W) <- as.character(1:nrow(W))
y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))

# Create model
data <- data.frame(y = y, mu = 1)
fm <- y ~ 0 + mu
X <- model.matrix(fm, data = data)

# Create framework for lists
setsGB <- list(A = colnames(W)) # gblup model
setsGF <- list(C1 = colnames(W)[1:1000], C2 = colnames(W)[1001:2000], C3 = colnames(W)[2000:10000]) # gfblup model
setsGT <- list(C1 = colnames(W)[1:10], C2 = colnames(W)[1001:1010], C3 = colnames(W)[1:10000]) # true model

# Compute G
G <- computeG(W = W)
GB <- lapply(setsGB, function(x) {computeG(W = W[, x])})
GF <- lapply(setsGF, function(x) {computeG(W = W[, x])})
GT <- lapply(setsGT, function(x) {computeG(W = W[, x])})

# REML analyses and single marker association test
fitGB <- greml(y = y, X = X, G = GB, verbose = TRUE)

#> [1] "Iteration:" "1"          "Theta:"     "2.31"       "1.99"      
#> [1] "Iteration:" "2"          "Theta:"     "4.63"       "3.03"      
#> [1] "Iteration:" "3"          "Theta:"     "8.91"       "3.5"       
#> [1] "Iteration:" "4"          "Theta:"     "14.98"      "2.31"      
#> [1] "Iteration:" "5"          "Theta:"     "19.7"       "0.47"      
#> [1] "Iteration:" "6"          "Theta:"     "20.83"      "0"         
#> [1] "Iteration:" "7"          "Theta:"     "20.86"      "0"         
#> [1] "Iteration:" "8"          "Theta:"     "20.86"      "0"         
#> [1] "Iteration:" "9"          "Theta:"     "20.86"      "0"         
#> [1] "Converged at Iteration:" "9"                      
#> [3] "Theta:"                  "20.86"                  
#> [5] "0"                      

maGB <- lmma(fit = fitGB, W = W)