Multi marker association analysis based on marker set tests

Genetic marker set tests based on summing the single genetic marker test statistics. The single marker test statistics can be obtained from GBLUP and GFBLUP model fits or from standard GWAS.

mma(stat = NULL, sets = NULL, ncores = 1, nperm = 1000,
  method = "sum")

Arguments

stat
vector of single marker statistics (e.g. marker effects, t-stat, p-values)

sets
list of marker sets - names corresponds to rownames in stat

ncores
number of cores

nperm
number of permutations

method
including sum, cvat, hyperG, score

W
matrix of centered and scaled genotypes (used if method = cvat or score)

fit
is the fit object obtained from a linear mixed model fit using the greml function

g
vector (or list) of genetic effects obtained from a linear mixed model fit (GBLUP of GFBLUP)

s
vector (or list) of single marker effects obtained from a linear mixed model fit (GBLUP of GFBLUP)

threshold
used if method = hyperG

Value

Returns a dataframe including

setT

marker set test statistics

nset

number of markers in the set

p

p-value for marker set

Details

The sum test is powerful if the genomic feature harbors many genetic markers having small to moderate effects. The distribution of this test statistic under the null hypothesis (associated markers are picked at random from the total number of tested genetic markers) is difficult to describe in terms of exact or approximate distributions, and an empirical distribution is required. Genetic marker set tests based on the covariance statistics for a set of genetic markers. The covariance test statistic is derived from a GBLUP (or GFBLUP) model fit. It is a measure of covariance between the total genomic effect for all markers and the genomic effect for the genetic markers in the genomic feature. It also relates to the explained sums of squares for the genetic markers. The distribution of these test statistics under the null hypothesis is difficult to describe in terms of exact or approximate distributions, and an empirical distribution is required. Genetic marker set tests based on the hyperG test statistics for a set of genetic markers. The hyperG marker set test tests a predefined set of markers (i.e. those within a particular genomic feature) for an association with the trait phenotype. Under the null hypothesis (associated markers are picked at random from the total number of tested genetic markers) it is assumed that the observed count statistic is a realization from a hypergeometric distribution. This hypothesis can be formulated and tested in a number of ways. Here we consider a test statistic based on counting the number of genetic markers in the feature that are associated to trait phenotype. A test based on the count test statistic is likely to have high power to detect association if the genomic feature harbours genetic markers with large effects.

References

Rohde, P. D., Demontis, D., Cuyabano, B. C. D., Boerglum, A. D., & Soerensen, P. (2016). Covariance Association Test (CVAT) Identifies Genetic Markers Associated with Schizophrenia in Functionally Associated Biological Processes. Genetics, 203(4), 1901-1913.

Rohde, P. D., Edwards S. M., Sarup P., Soerensen, P. (August, 2014). Gene-based Association Approach Identify Genes Across Stress Traits in Fruit Flies. Poster presented at the 10th World Congress of Genetics Applied to Livestock Production (WCGALP), Vancouver, Canada.

Examples

 # Simulate data
 W <- matrix(rnorm(20000000), ncol = 10000)
 colnames(W) <- as.character(1:ncol(W))
 rownames(W) <- as.character(1:nrow(W))
 y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))

 # Create model
 data <- data.frame(y = y, mu = 1)
 fm <- y ~ 0 + mu
 X <- model.matrix(fm, data = data)

 # Compute GRM
 GRM <- computeGRM(W = W)

 # REML analyses and single marker association test
 fit <- greml(y = y, X = X, GRM = list(GRM), verbose = TRUE)

#> [1] "Iteration:" "1"          "Theta:"     "2.25"       "2.01"      
#> [1] "Iteration:" "2"          "Theta:"     "4.4"        "3.2"       
#> [1] "Iteration:" "3"          "Theta:"     "8.2"        "4.12"      
#> [1] "Iteration:" "4"          "Theta:"     "13.45"      "3.69"      
#> [1] "Iteration:" "5"          "Theta:"     "17.53"      "2.36"      
#> [1] "Iteration:" "6"          "Theta:"     "18.59"      "1.87"      
#> [1] "Iteration:" "7"          "Theta:"     "18.64"      "1.84"      
#> [1] "Iteration:" "8"          "Theta:"     "18.64"      "1.84"      
#> [1] "Iteration:" "9"          "Theta:"     "18.64"      "1.84"      
#> [1] "Iteration:" "10"         "Theta:"     "18.64"      "1.84"      
#> [1] "Converged at Iteration:" "10"                     
#> [3] "Theta:"                  "18.64"                  
#> [5] "1.84"                   

 ma <- mlma(fit = fit, W = W)
 ma <- as.matrix(ma)

 f <- factor(rep(1:100,each=100), levels=1:100)
 sets <- split(as.character(1:10000),f=f)

# Set test based on sums 
res <- mma(stat = ma[,"stat"]**2, sets = sets, method = "sum", nperm = 10000)

#> Error in .Fortran("psets", m = as.integer(m), stat = as.double(stat),     nsets = as.integer(nsets), setstat = as.double(setstat),     msets = as.integer(msets), p = as.integer(rep(0, nsets)),     np = as.integer(np), ncores = as.integer(ncores), PACKAGE = "qgg"): "psets" not available for .Fortran() for package "qgg"

# Set test based on hyperG 
res <- mma(stat = ma[,"p"], sets = sets, method = "hyperG", threshold = 0.05)

#> Error in mma(stat = ma[, "p"], sets = sets, method = "hyperG", threshold = 0.05): unused argument (threshold = 0.05)