Genomic Feature Model analyses implemented using Restriced Likelihood Methods in DMU

Genomic Feature Best Linear Unbiased Prediction models implemented using REML.

remlDMU(fm = NULL, vfm = NULL, GRMlist = NULL, restrict = NULL,
  data = NULL, validate = NULL, bin = NULL)

Arguments

fm
a formula with model statement for the linear mixed model

GRMlist
a list of relationship / correlation matrices corresponding to random effects specified in vfm

data
a data frame containing the phenotypic observations and fixed factors specified in the model statements

validate
a matrix or a list with the ids of validation sets corresponding to the rows in data

bin
is the directory for DMU binaries (dmu1 and dmuai1)

Value

Returns results in a list structure including

f

list of predicted random effects

sigma

estimated variance components

asd

asymptotic standard deviation for the estimated variance components

fitted

fitted values from linear mixed model fit

residuals

residuals from linear mixed model fit

Details

The models are implemented using restricted maximum likelihood methods. Variance components estimated using REML and predictions are based on MME. Predicted random effects and single marker effects and statistics can be obtained. Cross validation procedures for assessment of prediction accuracy and model selection. This is an interface to be used for DMU.

References

Edwards, S. M., Sørensen, I. F., Sarup, P., Mackay, T. F., & Sørensen, P. (2016). Genomic prediction for quantitative traits is improved by mapping variants to gene ontology categories in Drosophila melanogaster. Genetics, 203(4), 1871-1883.

Examples

library(qgg)

setwd("C:/Users/sor/Dropbox/GFBLUP DGRP/scripts used in the analyses/work")

#> Error in setwd("C:/Users/sor/Dropbox/GFBLUP DGRP/scripts used in the analyses/work"): cannot change working directory

#bin <- "C:/Program Files (x86)/QGG-AU/DMUv6/R5.2-EM64T/bin"

# Simulate data
W <- matrix(rnorm(4000000), ncol = 10000)
  colnames(W) <- as.character(1:ncol(W))
  rownames(W) <- as.character(1:nrow(W))

G <- W %*% t(W) / ncol(W)

y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + 10 * rnorm(nrow(W))

data <- data.frame(f = factor(sample(1:2, nrow(W), replace = TRUE)), g = factor(1:nrow(W)), y = y)

fm <- y ~ f + (1 | g~G)
GRMlist <- list(G = G)

fit <- remlDMU(fm = fm, GRMlist = GRMlist, data = data)

#> Error in model$data$integers[model[[i]]$factors]: invalid subscript type 'list'

  str(fit)

#> Error in str(fit): object 'fit' not found