Genetic marker set tests based on sum statistics

Set test based on summing the single genetic marker test statistics. The sum test is powerful if the genomic feature harbors many genetic markers having small to moderate effects.

setTest(stat = NULL, W = NULL, sets = NULL, nperm = NULL,
  method = "sum", threshold = 0.05)

Arguments

stat
vector of single marker statistics (e.g. marker effects, t-stat, p-values)

W
matrix of centered and scaled genotypes (used if method = cvat or score)

sets
list of marker sets - names corresponds to rownames in stat

nperm
number of permutations

method
including sum, cvat, hyperG, score

threshold
used if method = hyperG

Value

Returns a dataframe including

setT

marker set test statistics

nset

number of markers in the set

p

p-value for marker set

Details

The singler marker test statistics can be obtained from GBLUP and GFBLUP model fits or from standard GWAS. The distribution of this test statistic under the null hypothesis (associated markers are picked at random from the total number of tested genetic markers) is difficult to describe in terms of exact or approximate distributions, and an empirical distribution is required.

Examples

# Simulate data
W <- matrix(rnorm(2000000), ncol = 10000)
  colnames(W) <- as.character(1:ncol(W))
y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))

# REML analyses 
data <- data.frame(y = y, mu = 1)
fm <- y ~ mu
fit <- gfm(fm = fm, W = W, sets = list(colnames(W)), data = data)

#> start algorithm at 11.8153 11.8153 
#> 1 sigma = 11.81529 11.81529 resid llik = -413.1626 
#> 1 adjusted sigma = 11.7223 11.7223
#> 2 sigma = 24.12272 5.73551 resid llik = -412.8265 
#> 2 adjusted sigma = 18.95108 4.505881 delta.llik = 0.3360854 
#> 3 sigma = 25.94281 1.227899 resid llik = -412.743 
#> 3 adjusted sigma = 22.43548 1.061894 delta.llik = 0.08350128 
#> 4 sigma = 26.87045 0.02522451 resid llik = -412.7278 
#> 4 adjusted sigma = 23.49198 0.02205299 delta.llik = 0.01521621 
#> 5 sigma = 26.64627 2.061154e-09 resid llik = -412.7275 
#> 5 adjusted sigma = 23.51441 1.818897e-09 delta.llik = 0.000272052 
#> Warning: solution lies close to zero for some positive variance components, their standard errors may not be valid

fit$df <- 10
fit$p <- pt(fit$s / sqrt(fit$vs), df = fit$df, lower.tail = FALSE)

sets <- list(A = as.character(1:100), B = as.character(101:1000), C = as.character(1001:5000), D = as.character(5001:10000))

# Set test based on sums 
res <- setTest(stat = fit$s**2, sets = sets, method = "sum", nperm = 100)

# Set test based on cvat 
res <- setTest(stat = fit$s, W = W, sets = sets, method = "cvat", nperm = 100)

# Set test based on hyperG 
res <- setTest(stat = fit$p, sets = sets, method = "hyperG", threshold = 0.05)