Prepare genotype data for all statistical analyses (initial step)

All functions in qgg relies on a simple data infrastructure that takes five main input sources; phenotype data (y), covariate data (X), genotype data (G or Glist), a genomic relationship matrix (GRM or GRMlist) and genetic marker sets (sets).

The genotypes are stored in a matrix (n x m (individuals x markers)) in memory (G) or in a binary file on disk (Glist).

It is only for small data sets that the genotype matrix (G) can stored in memory. For large data sets the genotype matrix has to stored in a binary file on disk (Glist). Glist is as a list structure that contains information about the genotypes in the binary file.

The gprep function prepares the Glist, and is required for downstream analyses of large-scale genetic data. Typically, the Glist is prepared once, and saved as an *.Rdata-file.

The gprep function reads genotype information from binary PLINK files, and creates the Glist object that contains general information about the genotypes such as reference alleles, allele frequencies and missing genotypes, and construct a binary file on the disk that contains the genotypes as allele counts of the alternative allele (memory usage = (n x m)/4 bytes).

The Glist structure is used as input parameter for a number of qgg core functions including: 1) construction of genomic relationship matrices (grm), 2) estimating genomic parameters (greml), 3) single marker association analyses (lma or mlma), 4) gene set enrichment analyses (gsea), and 5) genomic prediction from genotypes and phenotypes (gsolve) or genotypes and summary statistics (gscore).

gprep(study = NULL, fnRAW = NULL, bedfiles = NULL, bimfiles = NULL,
  famfiles = NULL, ids = NULL, rsids = NULL, overwrite = FALSE,
  ncores = 1)

Arguments

study
name of the study

fnRAW
path and .raw filename of the binary file used for storing genotypes on the disk

bedfiles
vector of names for the PLINK bed-files

bimfiles
vector of names for the PLINK bim-files

famfiles
vector of names for the PLINK fam-files

ids
vector of individuals used in the study

rsids
vector of marker rsids used in the study

overwrite
logical if TRUE overwite binary genotype file

ncores
number of cores used to process the genotypes

Value

Returns a list structure with information about genotypes

Examples

# NOT RUN {
#Glist <- gprep( bedfiles, bimfiles, study, path, additional arguments...)
#W <- getW( Glist, ids, rsids, additional arguments...)
# }